MAJOR CHANGE: adding scripts and changing existing scripts to be manual, see website and changelog
