+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.0
+* Appium now supported with the help of WDIO - see [mobile steps](https://courgette-testing.com/mobile-step-definitions) and [getting started page](https://courgette-testing.com/getting-started) for docs
+* Postinstall script has been removed and scripts have been added for running manually see [getting started page](https://courgette-testing.com/getting-started)
+
 ## 2.19.0
 * added compression for screenshot pngs
