+++ package.json
@@ -8,7 +8,6 @@
     "ci": "npm run install-chromedriver-ci && npm run start & npm run lint && npm run uitests",
     "eject": "react-scripts eject",
     "lint": "eslint '**/*.js'",
-    "postinstall": "node scripts/setup.js",
     "build-snippets:atom": "node scripts/generateSnippetsAtom.js --genFiles",
     "build-snippets:sublime": "node scripts/generateSnippetsSublime.js --genFiles",
     "build-snippets:vscode": "node scripts/generateSnippetsVSCode.js --genFiles",
@@ -29,6 +28,13 @@
   "bin": {
     "courgette": "./index.js",
     "ct": "./index.js",
+    "ct-install-firefox-driver": "./scripts/installFFDriver.js",
+    "ct-install-sample-conf": "./scripts/installSampleConf.js",
+    "ct-install-sample-mobile-conf": "./scripts/installSampleMobileConf.js",
+    "ct-install-sample-tests": "./scripts/installSampleTests.js",
+    "ct-package-scripts": "./scripts/setupPackageJsonScripts.js",
+    "ct-package-scripts-mobile": "./scripts/setupPackageJsonScriptsMobile.js",
+    "ct-setup-snippets": "./scripts/setupSnippets.js",
     "uitests": "./index.js"
   },
   "repository": {
