+++ scripts/util.js
@@ -1,0 +1,37 @@
+#! /usr/bin/env node
+const childProcess = require('child_process');
+
+function runScript(scriptPath, args) {
+  let invoked = false;
+
+  const process = childProcess.fork(scriptPath, args);
+
+  return new Promise((resolve, reject) => {
+    process.on('error', (err) => {
+      if (invoked) {
+        reject(new Error('error already invoked'));
+        return;
+      }
+      invoked = true;
+      reject(err);
+    });
+
+    process.on('exit', (code) => {
+      if (invoked) {
+        reject(new Error('exit already invoked'));
+        return;
+      }
+      invoked = true;
+
+      if (code === 0) {
+        resolve();
+      } else {
+        reject(new Error(`exit code ${code}`));
+      }
+    });
+  });
+}
+
+module.exports = {
+  runScript,
+}
