+++ uiTestHelpers/mobileStepDefinitions/checks/checkAttribute.js
@@ -7,7 +7,7 @@ module.exports = async function checkAttribute(locatorKey, expectedAttribute, ex
   if (process.env.DEBUG) {
     console.log('In file:', path.basename(__filename), el);
   }
-  
+
   const attrVal = await el.getAttribute(expectedAttribute);
 
   return expect(attrVal).to.equal(expectedValue);
